refactor(client): refactor challs frontend
Resolves https://github.com/redpwn/rCTF/issues/54
