+++ client/src/index.js
@@ -10,7 +10,7 @@ import Home from './routes/home'
 import Registration from './routes/registration'
 import Login from './routes/login'
 import Profile from './routes/profile'
-import Challenges from './routes/challenges'
+import Challenges from './routes/challs'
 import Scoreboard from './routes/scoreboard'
 import Error from './routes/error'
 import Sponsors from './routes/sponsors'
